Add main (tcp) reader algorithm
The function 'handle_msg_reader' sequentially cuts the
serde-serialised and delimited Messages out of the reader, unwraps
and deserialises them and finally sends them through the
transmitter.

The messages are separated from each others with the aid of some
special delimiter chars, which are not insertable as username or as
content of a Message. Besides, they are different from '{' and '}',
i.e. the parenthesis used in JSON format.