Add error handling for Message + '#' decorations in `lib.rs`
The enum TextValidityError has been defined to validate the text of
the username and of the content of messages and is now used in
Message. The tests for it passed successfully. Besides the fn
`Message::paket` now return also a Result.

`cargo check --lib` does not return errors or warnings.

`client` and `server` have not been adapted to the new change yet!

Finally the decorative frames of '#' have been added.