Refactor DELIMITERS, PAKET and related + ensure tests fragility
* The delimiters were specific chars with 1 byte in their byte
version, transposed then as bytes. This had two problems:

1. Such bytes could be present in the text, which had to be checked
for them.

2. One had to choose chars which are 1 byte long, and the program
needed to be completed with a check on such choices.

This was addressed by choosing instead of these, two bytes which do
not appear in the UTF8 encoding, and by chaging the definition of
paket: now a paket is always a Message which underwent
serialization and transposition into bytes, and then is sorrounded
by the two delimiters. Notice that a pleasant side effect is that
the TextValidityError now comes only for the length of the username
and the content, and not for the validity of the text.

In consequence of this also the function 'handle_msgs_reader' was
renamed to 'message_depaketer' and consequently the tests naming.

* In the tests '?' was substituted with '.unwrap()' to ensure that
the tests stay fragile.