Adapt 'server.rs' + Add error_handling in it
* Change the usage of 'pakets_extractor' accordingly.

* Implement error-handling (create Error types for each async fn).
If the async fn B with error-type BErr spawns a task to run the
async fn A with error-type AErr, then AErr is converted in the
tokio_task into the BErr and then bubbledup.

Notice that we use the TaskTracker struct to spawn tasks, a part
from the case of the server_manager, where the JoinSet function is
used instead. This is becasue TaskTracker works better to manage
thousands of twin tasks, while TaskTracker is more apt to manage a
small amount of essentialy different tasks (and thus with different
error-types).