Reduce error-fragility for Message
* A new struct Username was created, which checks fro the length of
the username on its creation.

* Message::new is now private and takes Username as first argument
(already checked) and content as second argument. The max length of
a message has been redesigned to be much larger.

* A new 'impl' section has been created for Message, to allow the
client and server binaries to craft common messages as '{username}
joined the chat' and similar, in a canonical way and without
handling a 'Message::new' Result.