Reduce error-fragility for the message_depaketer and rename it 'paket_extractor'
* Such function now just divide the incoming data in pakets and
send them through the channel.

* Besides it has been adapted to teh new Message::MAX_CONTENT_LEN.
This is in fact so long that one cannot assume anymore certain
assumptions on the buffer dimension. In particular a message could
span over many (more than 2) buffers, which means that we need to
address also the problem that the vector 'previous_fragment' grows
over the maximum allowed length for a Vec in rust.