Update the unit-tests of async_functions with graceful-shutdown
Now the test does not need anymore that the developer presses
CTRL+C at the end of it, and checks also that the functions do
return exactly when we want.

Notice that 'test_pakets_extractor' sometimes fails because the
pakets_extractor finds a too long paket. When this happens the
paket is dropped, and thus not all messages arrive. The consequence
is that the loop of the message_receiver never finishes. Thus that
test checks also that all messages have arrived within 100
milliseconds.

Such error happens rarely, and only when the messages
length is the maximum amount available, thus it was plausibile to
think that this happens because of a small error in the byte being
read rather than because of the code.