tmp - Try fixup for async_function
I manually tested the client-server apparatus, and when the client
was interrupting the connection the server was not signalling that.
I conjecture that this might be related to the function not exiting
at the right moment.

Btw, I think that the two async functions I wrote are definitely
oto complicated, I should refactor them in a commit and decompose
them in smaller pieces. If i leave them like they are now It will
be very very difficult to review them.